Basic Webworker Environment Support
Updates to caman.coffee checking if we're in a worker and avoiding use
of non-worker-friendly stuff like canvases.

Util.coffee changed a reference from window to root to stop error
squawking.

In a worker, caman takes an imageData array and processes it. This data
can be gotten with a user defined function that must be titled
`exposeImageData(data, width, height)`.